misc.hh: No must be no default implementation for the CompileTimeChecker.
albertgrid.hh, .cc: Checked the implementation of the grid interface
and implemented the method father_relative_local but not check because no
test available. Furthermore fixed a bug in the method global and local of
AlbertGridElement.

lagrangebase.hh: The BaseFunctions for lagrange 1 on triangle changed back
to the old implementation.

[[Imported from SVN: r138]]