Changed the interface methods of the LocalOperators, because the old made not so much sence. Now there are five methods:
prepareGlobal( Param1 & , Param2 &) which is called with a Param1 and
  Param2 Parameter class which is specified by the user and this method is
  called before the grid wlakthrough starts.

then on each entity:

  prepareLocal ( en1 ) or preapreLocal(en1, en2) ( father , son )
  applyLocal   ( en1 ) ....
  finalizeLocal( en1 ) ....

after the grid walkthrough

  finalizeGlobal( )

The main changes are that now only in the preprareGlobal method we pass the
parameters and store them in the whatever class.
Furthermore this method is to combine operations on entitys, and we dont
want to know for which operators ( Operators on DiscreteFunctions or
AdaptionOperators or something else).
So the interface could be more gernal, but C++ has its restricts.

So far working on a general combination procedure.

[[Imported from SVN: r435]]