Updated IndexSet.
I already changed YaspGrid and SGrid but not the other grids.

Added begin/end methods and export Iterator type.

This required changes in the GridTraits (there are two additional template parameters),
see SGrid as an example.

The geomTypes method in IndexSet (and implementations) now has a parameter codim because
it is necessary to know for each codim what the types in the grid are. Single
element type grids do not have to change the implementation, but UGGrid must return the
appropriate information.

[[Imported from SVN: r2909]]