better implementation of geometryInFather.
separation of LeafIterator and LevelIterator moved from TreeIterator to
EntityPointer, because entities alos have to know, where they come from.
This information is stored be getNewEntity is called.
This Information is needed by the IntersectionIterator. Now the iterator
does different thing when started from and LevelIterator Entity.

The method getNewEntity is specialised now for entities of codim 0,
should be much faster now.

Also bug fix in intersection iterator, when setting up the neighbor in 3d,
the face twist has to be calculated. The neighbors are set up right now.

[[Imported from SVN: r3285]]