forgot to say:
Entity,Geometry, and IntersectionIterator have a method getRealImp now
that returns a reference to real implementation.

This method can only be accessed by the GridDefaultImplementation.
Therefore all GridImplementations the inherit from
GridDefaultImplementation can use this a method on
GridDefaultImplementation (getRealImplementation) that gives access to the
real implementation of this wrapper classes.

[[Imported from SVN: r3807]]