Removed the specialization of Geometry for dim=0.  That way, vertices have the full set of methods as the higher-dimensional geometries.
It would be possible to provide default implementations for the new
dim=0 methods in the wrapper class.  This would not lead to any
efficiency problems, as any if(mydim==0) can get optimized away.
However, the different grid implementations still need to have
the functions in order to have the code compile.  So for now I
didn't include default implementations.

[[Imported from SVN: r4079]]