Changed README to honor dunecontrol as well as the new licence.
WARNING:

Currently COPYING contains the GPL without the exception. The runtime
exception is just mentioned in the README. It is  not clear to me
whether it should be included into the COPYING file. In libstdc++ it
is not included there but in all source files.

Sven: please audit the licence again.

[[Imported from SVN: r4867]]