Add a mechanism for the transition from FieldVector::size as an enum member to FieldVector::size() as a method.
With this patch applied FieldVector::size remains an enum, hence all your
code compiles as before.  However you get a compiler warning explaining
the issue and suggesting to configure with --enable-fieldvector-size-is-method.
If you do so, then FieldVector::size becomes a method, and you may have
to go out and adjust your code.  The configure flag is transitional and
will be removed after dune 2.2 is out.

Remember that there is no smoother way to do the transition, as you can't
have FieldVector::size to be both an enum value and a method at the same
time.

This fixes FlySpray issue 819.


[[Imported from SVN: r6486]]