Allow the user to override whitespace hook setting in dune.module
This patch makes it possible to bypass installation of the whitespace hook by specifying

VCSETUP_FLAGS="DISABLEWHITESPACEHOOK=1"

in the user's options. If dunecontrol encounters that flag, it will not attempt to install
the hook in any module, uninstall the hook if it has already been installed and output a
warning telling the user that he'll have to take care of whitespace by himself.

[[Imported from SVN: r7471]]