[dunecontrol] Fix parsing of Git configuration settings file
dunecontrol fails to properly process Git settings where the value contains
spaces. It was supposed to handle quoting the same way the shell does, but
instead arguments were split at each space and quotes kept verbatim in the
split arguments.

The reason is the order in which the shell performs variable substitution and
tokenizing, preventing the line with the Git configuration entry from being
tokenized by the shell.

This patch fixes the problem by first parsing the line into an array, which is
then in turn expanded again with correct per-element quoting.

In addition, make COMMENT helper variables local while we're at it.