[CMake] Parse other module's dune.module to set variables for config.h
To set variables like DUNE_COMMON_REVISION the macro dune_module_information
is called. To surpress output the macro gets an additional second
argument to QUIET the output. To set the PACKAGE_* variables correctly
dune_module_information is called a second time for the current module
after it was called all the other modules.