[FixedBlockSizeRange] Provide direct access to the range of blocks corresponding to the scalar range
In addition to begin() and end(), there are now also methods begin_block() and
end_block() that hand out block indices. This functionality is useful if the
per-thread algorithms work in terms of blocks (as the upcoming SIMD kernels do).

You have to be careful about padding, though: While end() will never let you iterate
past the actual end of the range, end_block() will point to the end of the last,
potentially not completely filled block. So to efficiently use the blocked interface,
you have to make sure to pad your datastructures to the block size.