[TBB] Extend autoconf test to extract alignment of tbb::cache_aligned_allocator
We need the guaranteed alignment of tbb::cache_aligned_allocator as a compile
time constant for the SIMD kernels, so this test tries to extract it from the
TBB sources. This will of course fail if you use the binary-only version of TBB
from Intel. Unfortunately, the only alternative is to call a function for the value,
but that fails when we are cross-compiling (think Xeon Phi).

So this test still needs to be fleshed out, but for now it works.