[Memory] Add static algorithm to verify that a number is 2^n and calculate n
In some of the containers using the new block-based kernel algorithms,
it is often necessary to convert scalar, flat coordinates into a block
number and an offset. Usually, this operation requires a division and
a mod operation, but as most kernels require the block size to be a
power of two, we can reduce the cost by instead doing a shift and a mask
operation.

This patch can be used to obtain the correct shift value from a given
block size and also to make sure that the block size actually is a power
of 2.