[Threads] Add chunk_size parameter to fixed_block_size_range
The fixed_block_size_range operates on blocks and will always guarantee
that range division happens on a block boundary. However, sometimes
alignment constraints require even stricter division requirements, only
allowing the split to happen on a multiple of the block size.

This patch adds an additional parameter chunk_size to the range. This
parameter determines how many blocks will be grouped together to a chunk
before doing subdivion on those chunks. The parameter defaults to
1 (i.e. split on any block boundary).