[Kernel] Add masked dot product kernel for vectors
The masked dot product takes a third vector containing masking
information and a functor which is passed matching elements from all
three vectors. The return value of the functor is multiplied with the
dot product of the two data vectors.

As the functor will have to be inlined for decent performance, there are
no precompiled library versions of this kernel.