[CMake][bugfix] Remove CMAKE_REQUIRED* from FindUMFPack.cmake.
The removed lines would be necessary if the included headers
or libraries would be tested. As this is not the case, they can
be safely removed.
The way they were used, every test tried to link to the libraries
amd and umfpack causing every test to fail if the libraries are
not present. The changed declaration of CMAKE_REQUIRED_* should
be encapsulated together with the test inside
  include(CMakePushCheckState)
  cmake_push_check_state()
and
  cmake_pop_check_state()
to ensure that the definitions do not leak the test.