Fixes setting of module path for installed package.
Previously the module path was a relative to the install
directory. Because of this dependant modules did not find
the macros of dune-common and cmake configure did not succeed.
With this the CMake package configuration file contains the
full path and the macros are found (even if there is no local
dune-common lying around).
(cherry picked from commit db6f61c24754c42e191a6ed760772aedc860d2df)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>