Fixes setting of module path for installed package.
Previously the module path was a relative to the install
directory. Because of this dependant modules did not find
the macros of dune-common and cmake configure did not succeed.
With this the CMake package configuration file contains the
full path and the macros are found (even if there is no local
dune-common lying around).