[release,parallel] Allow copying of RemoteIndexListModifier.
Previously, copying a RemoteIndexListModifier resulted in a double
free of a pointer to a list of global indices. There was no need to
use a pointer to this, though. After this patch we store the list directly
by value. Thus the default copy constructor works but makes a rather cheap copy
of a (most often empty) singly linked list and we get rid off the non-default
destructor.
(cherry picked from commit d561f032c6176668a328fefb200f772e7912936b)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>