[buildsystem] distribute cxx11_constexpr.m4
The check for constexpr was not copied into the distribution.
Unfortunately, the autotools will only warn about a missing macro, but
give valid results otherwise. This patch adds the cxx11_constexpr.m4 to
Makefile.am, so that it is distributed on "make dist".
(cherry picked from commit 3887555aa286141b1476fde7ad98ebf0217e2805)

Signed-off-by: Christoph Grüninger <gruenich@dune-project.org>