[release] Honor USE_CMAKE environment variable and allow deactivating it on the command line.
With this patch the USE_CMAKE environment variable is queried whether or not CMake
should be used. This can be overwritten in an options file or on the command line
using either the --cmake or --no-cmakr switch.
(cherry picked from commit cba27f049f5b5bc8b92ea86764ed1b1ff06fd9b0)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>