[Release][MPI] Add check for MPI standard version and warn if installed version is < 2.1
As decided in the doodle for FS#1395, really old MPI versions will be
deprecated in the 2.3 release.

This patch adds an additional check to the MPI tests for autotools and
cmake that makes sure the user has an MPI which is at least compliant
with MPI-2.1. If the user's version is older, a warning is generated.

The test uses version macros instead of the more standard runtime
MPI version query interface to avoid problems when cross-compiling.
(cherry picked from commit ac6885f8af7c6b63e99f3285ef2a98d2c5523637)

Signed-off-by: Christoph Grüninger <gruenich@dune-project.org>