[autotools,bugfix] Fixes documentation building in tarballs.
When unpacking the tarballs, the time stamps are preserved, but the
access time of the files might be set differently. In addition during
configure Doxyfile is always created. This resulted in some files being
newer than the doxygen-tag in the tarball and triggered building the
documentation. This patch fixes the problem for me by only checking the
modification time and excluding Doxyfile.
(cherry picked from commit 127aa0a3663c4df84646d9e9a7f55d4170638ee4)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>