[Doxygen][release] Set INCLUDE_GRAPH and INCLUDED_BY_GRAPH to NO.
With openSuse 12.3 (Doxygen 1.8.5) the generated documentation contains
lots of huge graphs which show the include dependency graph of a file.
This lead to large tarballs of 155 MiB compared to 50 MiB with this
patch.
With Debian Wheezy (Doxygen 1.8.1.2) the tarballs did not blow up in
file size. Scanning the Doxygen change log did not enlighten me and
I remain puzzled why the graph were never built.

INCLUDED_BY_GRAPH is changed to prevent the same behvior for future
releases. We currently do not include these graphs and this change
makes the configureation consistent with the result.