[release,bugfix] Makes PKG_CONFIG_PATH honor all (lib lib/multiarch).
Previously some of the pathes (e.g the multiarch ones) took precedence
over the others. But as with same some of the pkgconfig files are installed
below lib/<multiarch> and others (if there is not object library) are not.
With this patch all possible install locations of pkgconfig files are added
to the path if they exist.
(cherry picked from commit 0aabf5fa839d50b086cddaf7fa741726ed8fecb9)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>