[parallel,bugfix] don't create empty MessageInformation
when we try to send data, it might happen, that for some reason we don't want to any data.
This happens, when the dynamic size of our CommPolicy is 0. In that case we must not add a
MessageInformation, otherwise we get an MPI error.(cherry picked from commit 82609fbcd3b8a49cdc5dc8cd640f55609622586e)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>