[bugfix,cmake] Always read dune.module after finding a module to extract dependencies.
Previously we already processed dune.module but in a much later stage during
configure. Dependencies are contained in dune.module. Therefore there is no
need to replicate them in the CMake package configuration files. With this
patch we now always process dune.module directly after finding the DUNE module.
In addition we set ${DUNE_MOD_NAME}_(DEPENDS|SUGGESTS} in the CMake environment.
As with this change DUNE_MOD_NAME will usually be set to to dune-common, we
also changed to using ProjectName and ProjectVersion (the canonical CMake names)
to refer to the current module and version.