[cmake,bugfix] Fixes issues in dependency tracking.
In this patch we rename dune_process_depedency_tree to
dune_process_depedency_macros because it actually just
include the macros of all modules that we depend on.
I addition we fix a bug that resulted in not executing the module's
own macro if there was only one dependency and now always process
it here.