[poolallocator,release] Use a non-static memory pool.
Due to its desgin of using one static memory pool. The allocator
could not be used in different threads at the same time without
side effects. With this revision each instance will hold its own private
instance of the pool.

There should be no issues with this patch and poolallocator seems to be
mostly used in AMG anyway.