[bugfix][release] Added missing copy constructor for Pool.
When copying PoolAllocators Bard experienced linking errors to the
copy constructor of Pool, which is deactivated. To fix this we added
the copy constructor. Note that it does not copy the pool such that
ownership of allocated memory chunks is clear.

Kudos go to Bard for noticing this and providing the fix.

In addition we removed the copy constructor of the specialization for
void as the default coy constructor suffices.