[Buildsystem] Require C++11-compliant compiler and clean up associated tests
As we have bumped our compiler requirements to GCC 4.4+, we can now
require the compiler to support C++11.

This patch updates both the autotools and the CMake tests to produce an
error if the compiler doesn't support either -std=c++11 or -std=c++0x.

It also renames the tests from ..0X to ..11 and renames the configure
options from --disable-gxx0xcheck to --disable-cxx11check and removes a
couple of redundant calls to the autotools macro.