[dunecontrol] Fixes error with installed old DUNE modules
If duncontrol detects an installed dune-common module, then it always
assumed that it contains a lib/dunecontrol
subdirectory. Unfortunately, this does not hold for older dune-common
modules and dunecontrol will error out later with a message like

/usr/local/lib/dunecontrol: file not found
Execution of dunecontrol terminated due to errors!

With this commit we fix this by actually checking whether the
lib/dunecontrol subdirectoy exists before adding it to
DUNE_CONTROL_PATH.

Thanks a lot to wuhao sife for reporting this problem.