[bugfix] include memory header to make make_shared available.
According to the standard make_shared is defined in the memory header.
While we did include the header when testing for make_shared support, we
forgot to include it in dune/common/shared_ptr.hh. This patch now uses the
same include statement for memory in dune/common/shared_ptr.h as in the
m4 test.

This problem does only exist in the release as Christoph already fixed in the
trunk while stripping our own make_shared implementation.

Kudos to Tatiana and Eike for finding this bug!