[MPI] Make parallel version CollectiveCommunication default-constructible
In order to simplify fallback handling between code with and without MPI
support, this patch makes the specialization of CollectiveCommunication
for MPI communicators default-constructible by defaulting the underlying
MPI communicator to MPI_COMM_WORLD, which should be reasonable for most
scenarios.

As the non-specialized version for sequential code is already
default-constructible, this allows grid constructors to accept a
CollectiveCommunication object regardless of whether MPI support is
enabled or not and to provide a default value for this parameter.

This patch resolves FS#1497.