[test] [CMake] Make debugallocator tests pass with ctest
testdebugallocator_fail{1..5} are a very special sort of tests,
as in order to pass they are expected to fail with a signal like
SIGABRT or SIGSEGV. While autotools, dont distinguish much
about how a test failed, if it is supposed to fail, ctest
handles signals from the test and distinguishes this from
normal termination with nonzero exit code.

This commit introduces a signal handling in the test itself,
that causes the program to terminate "normally" with a nonzero exit
code instead of passing the signal to the fathering process
(ctest). The signal, which is expected to be received can be set
from CMakeLists.txt via a preprocessor variable.