Implement a closed-form formula for the eigenvalues of a symmetric 3x3 matrix
I took the formula from

 http://en.wikipedia.org/wiki/Eigenvalue_algorithm  (retrieved lated August 2014).

Wikipedia claims to have taken it from

 Smith, Oliver K. (April 1961), Eigenvalues of a symmetric 3 × 3 matrix,
 Communications of the ACM 4 (4): 168, doi:10.1145/355578.366316

I did not check stability of efficiency of this, but no problems related to
either are mentioned in the literature.  The main reason I need the closed-form
expression is to be able to pipe it into an algorithmic differentiation system.