[parametertree][locale][fs1528] Check ParameterTree::get() with a locale using "," as the decimal seperator.
ParameterTrees are usually read from configuration files.  The format of the
configuration file should not depend on the locale.

For the commandline parser it would likewise be surprising if "progname -param
0.5" stopped working in a german locale.

Not long ago these kinds of errors would not even generate a diagnosis due to
FS#1527.  However, the locale of the program is "C" by default, so probably
not many poeple are actually affected by this problem.  It can happen however
that other libraries linked into the same program set the locale from the
environment.