[cmake] Allow for custom sections in automatically generated package configuration files.
In dune-grid we want to export the path to the grid example files. To do this
without a handwritten package configuration file, We add a custom section
to it that can be filled by the modules using the variable DUNE_CUSTOM_PKG_CONFIG_SECTION
before finalizing the cmake build system. For dune-grid we export the variable
DUNE_MODULE_SRC_DOCDIR that points to the installed documentation and the path to
the source files of the documenation in the installed and not-installed module, respectively.