[CMake] Introduce a check, whether the user runs into FS1532
With a compiler name given to CMAKE_CXX_COMPILER (instead of a path),
the name is correctly expanded on the first run of cmake, but as soon
as the buil didrectory is existent, it is for obscure reasons expanded
into ${CMAKE_BINARY_DIR}/.... To avoid having users run into this bug,
the expanded compiler variable is matched with the build directory prefix.
In that case, an error is issued.