[CMake] Add macro dune_symlink_to_source_tree
Add a symlink called src_dir to all directories in the build tree.
That symlink points to the corresponding directory in the source tree.
Call the macro from the toplevel CMakeLists.txt file of your project.
You can also call it from some other directory, creating only symlinks
in that directory and all directories below. A warning is issued on
Windows systems.