[Cmake][Bugfix] Fix FindMETIS by adding -lm to the link flags
It needs to be linked in some cases. The way of checking whether
it is needed is taken from Find{BLAS,LAPACK}.cmake . This fix was
needed as part of a larger fix for the test of parallel alugrid.

(cherry picked from commit c92679d89f835868d7a6a6f4ea3ba524d0ae0c58)